requirements.txt
boto3==1.18.64
Django==3.2.8
psutil==5.8.0
pymemcache==3.5.0
pymongo[srv]
CherryPy==18.6.1
Routes==2.5.1
django-dotenv==1.4.2
htmlmin==0.1.12
gunicorn
psycopg2
sitemap42==2021.1.17
progress==1.6